Fixed Bend Time collision issue
Essentially, when time was 'frozen' the collision calculations with physics bodies, i.e. spheres, would be multiplied by a stupid amount.

To fix this, turned off simulate physics if a physics body was too close to the player and back on when (1) the timer ended or (2) the player manually ended bend time.